Define s32 and u32 to match __s32 and __u32.

+++ src/base/lib/linux-headers/asm-generic/int-ll64.h
@@ -43,8 +43,8 @@ typedef unsigned char u8;
 typedef signed short s16;
 typedef unsigned short u16;
 
-typedef signed int s32;
-typedef unsigned int u32;
+typedef signed /* int modified to match stdint.h */ long s32;
+typedef unsigned /* int modified to match stdint.h */ long u32;
 
 typedef signed long long s64;
 typedef unsigned long long u64;
